+++ websocket/src/main/kotlin/lt/saltyjuice/dragas/chatty/v3/websocket/route/WebSocketRoute.kt
@@ -1,5 +1,7 @@
 package lt.saltyjuice.dragas.chatty.v3.websocket.route
 
+import lt.saltyjuice.dragas.chatty.v3.core.middleware.AfterMiddleware
+import lt.saltyjuice.dragas.chatty.v3.core.middleware.BeforeMiddleware
 import lt.saltyjuice.dragas.chatty.v3.core.route.Route
 
 
@@ -18,9 +20,26 @@ open class WebSocketRoute<Request, Response> : Route<Request, Response>()
      */
     open class WebSocketRouteBuilder<Request, Response> : Route.Builder<Request, Response>()
     {
-        /**
-         * Implementations should return a raw route object which is later used in [adapt] to add all the callbacks, middlewares, etc.
-         */
+        override fun before(clazz: Class<BeforeMiddleware<Request>>): WebSocketRouteBuilder<Request, Response>
+        {
+            return super.before(clazz) as WebSocketRouteBuilder
+        }
+
+        override fun after(clazz: Class<AfterMiddleware<Response>>): WebSocketRouteBuilder<Request, Response>
+        {
+            return super.after(clazz) as WebSocketRouteBuilder
+        }
+
+        override fun description(string: String): WebSocketRouteBuilder<Request, Response>
+        {
+            return super.description(string) as WebSocketRouteBuilder
+        }
+
+        override fun adapt(route: Route<Request, Response>): WebSocketRoute<Request, Response>
+        {
+            return super.adapt(route) as WebSocketRoute
+        }
+
         override fun returnableRoute(): WebSocketRoute<Request, Response>
         {
             return WebSocketRoute()